backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
httpx>=0.27.0
google-api-python-client>=2.120.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
duckdb>=0.10.0
polars>=0.20.0
fastexcel>=0.9.0
python-multipart>=0.0.9
pydantic-settings>=2.2.0
